Tunning the build, Tunnung engines adding samples, Removing log4j and commons logging from distrib and jars moving the tld
git-svn-id: http://svn.code.sf.net/p/jcaptcha/code/trunk@180 eaf2ec09-4c2e-0410-84d0-fe0bfca4863a

+++ jcaptcha/engine/src/java/com/octo/captcha/engine/image/fisheye/SimpleFishEyeEngine.java
@@ -108,7 +108,7 @@ public class SimpleFishEyeEngine extends ListImageCaptchaEngine
 
         //add background from files
         BackgroundGenerator generator = new FileReaderRandomBackgroundGenerator(
-                new Integer(300), new Integer(300),
+                new Integer(250), new Integer(250),
                 "./fisheyebackgrounds");
         addFactory(
                 new FishEyeFactory(generator, sphereDef, new Integer(30),
